+++ CHANGELOG.rst
@@ -4,9 +4,13 @@ Change Log
 Unreleased
 ----------
 
-- Add ``module-toctree`` and ``package-toctree`` directives.
+- New ``remote-code-block``, which works like the ``literalinclude`` directive, but allows you to include content from a URL over the web.
+  You can use this directive after adding ``documenteer.sphinxext`` to the extensions list in a project's ``conf.py``.
+
+- New ``module-toctree`` and ``package-toctree`` directives.
   These create toctrees for modules and packages, respectively, in Stack documentation sites like pipelines.lsst.io.
   With these directives, we don't need to modify the ``index.rst`` file in https://github.com/lsst/pipelines_lsst_io each time new packages are added or removed.
+  You can use this directive after adding ``documenteer.sphinxext`` to the extensions list in a project's ``conf.py``.
 
 - New ``stack-docs`` command-line app.
   This replaces ``build-stack-docs``, and now provides a subcommand interface: ``stack-docs build`` and ``stack-docs clean``.
@@ -21,6 +25,8 @@ Unreleased
 
 - Recognize a new field in the ``metadata.yaml`` files of Sphinx technotes called ``exclude_patterns``.
   This is an array of file or directory paths that will be ignored by Sphinx during its build, as well as extensions like our ``get_project_content_commit_date`` for looking up commit date of content files.
+
+- Updated to Sphinx >1.7.0, <0.2.0.
 
 - Updated to lsst-sphinx-bootstrap-theme 0.3.x for pipelines docs.